+++ pojo/src/main/java/com/ybk/dto/match/EndMatchDTO.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EndMatchDTO {
-    private Long matchId;
-    private Long teamId;
+    private Integer matchId;
+    private Integer teamId;
 }
