+++ pojo/src/main/java/com/ybk/dto/match/MatchAModeDTO.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class MatchAModeDTO {
-    private Long matchAId; // 关联的matchA
+    private Integer matchAId; // 关联的matchA
     private String mode; // 比赛模式
     private LocalDateTime beginTime; // 比赛时间
     private Integer venueNumber; // 场地编号
