+++ server/src/main/java/com/ybk/service/RefereeService.java
@@ -13,7 +13,7 @@ public interface RefereeService {
 
     PageResult pageQuery(PageQueryDTO pageQueryDTO);
 
-    void passReferee(List<Long> ids);
+    void passReferee(List<Integer> ids);
 
     void save(RefereeDTO refereeDTO);
